feat: departamento, demanda e funcionario

+++ src/main/java/br/udesc/edu/demandatech/service/DepartamentoService.java
@@ -3,7 +3,9 @@ package br.udesc.edu.demandatech.service;
 import br.udesc.edu.demandatech.model.dto.criar.DepartamentoCriarDTO;
 import br.udesc.edu.demandatech.model.dto.editar.DepartamentoEditarDTO;
 import br.udesc.edu.demandatech.model.entity.Departamento;
+import br.udesc.edu.demandatech.model.entity.Funcionario;
 import br.udesc.edu.demandatech.model.exception.IdNotFound;
+import br.udesc.edu.demandatech.model.exception.PermissionDenied;
 import br.udesc.edu.demandatech.repository.DepartamentoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -17,14 +19,20 @@ import java.util.Optional;
 public class DepartamentoService {
     private DepartamentoRepository departamentoRepository;
 
-    public Departamento criar(DepartamentoCriarDTO departamentoCriarDTO) {
+    public Departamento criar(DepartamentoCriarDTO departamentoCriarDTO, Funcionario funcionario) {
+        if(!funcionario.getAdmin()) {
+            throw new PermissionDenied();
+        }
         Departamento departamento = new Departamento();
         BeanUtils.copyProperties(departamentoCriarDTO,departamento);
 
-        return departamento = departamentoRepository.save(departamento);
+        return departamentoRepository.save(departamento);
     }
 
-    public Departamento atualizar(Long id, DepartamentoEditarDTO departamentoEditarDTO) {
+    public Departamento atualizar(Long id, DepartamentoEditarDTO departamentoEditarDTO, Funcionario funcionario) {
+        if(!funcionario.getAdmin()) {
+            throw new PermissionDenied();
+        }
         Optional<Departamento> opcionalDepartamento = departamentoRepository.findById(id);
         if(opcionalDepartamento.isPresent()) {
             Departamento departamento = opcionalDepartamento.get();
@@ -38,7 +46,7 @@ public class DepartamentoService {
         return departamentoRepository.findAll();
     }
 
-    public Departamento buscarPorId(Departamento departamento, Long id) {
+    public Departamento buscarPorId(Long id) {
         Optional<Departamento> opcionalDepartamento = departamentoRepository.findById(id);
         if(opcionalDepartamento.isPresent()) {
             return opcionalDepartamento.get();
@@ -46,7 +54,18 @@ public class DepartamentoService {
         throw new IdNotFound("departamento", id);
     }
 
-    public void removerPorId(Long id){
+    public Departamento buscarPorName(String descricao) {
+        Optional<Departamento> opcionalDepartamento = departamentoRepository.getDepartamentoByDescricao(descricao);
+        if(opcionalDepartamento.isPresent()) {
+            return opcionalDepartamento.get();
+        }
+        throw new IdNotFound("departamento", id);
+    }
+
+    public void removerPorId(Long id, Funcionario funcionario) {
+        if(!funcionario.getAdmin()) {
+            throw new PermissionDenied();
+        }
         Optional<Departamento> optionalDepartamento = departamentoRepository.findById(id);
         if(optionalDepartamento.isPresent()) {
             departamentoRepository.deleteById(id);
